Crate rustc_smir

The WIP stable interface to rustc internals.

For more information see https://github.com/rust-lang/project-stable-mir

Note

This API is still completely unstable and subject to change.

Modules

rustc_internal

Module that implements the bridge between Stable MIR and internal compiler MIR.

rustc_smir 🔒

Module that implements what will become the rustc side of Stable MIR. This module is responsible for building Stable MIR components from internal components.

Macros

run

Instantiate and run the compiler with the provided arguments and callback.

run_with_tcx

Instantiate and run the compiler with the provided arguments and callback.