Module rustc_smir::rustc_internal

Module that implements the bridge between Stable MIR and internal compiler MIR.

For that, we define APIs that will temporarily be public to 3P that exposes rustc internal APIs until stable MIR is complete.

Structs

CrateNum

DefId

A DefId identifies a particular definition, by combining a crate index and a def index.

Functions

adt_def

alias_def

br_named_def

closure_def

crate_item

crate_num

fn_def

foreign_def

generator_def

impl_def

item_def_id

opaque 🔒

param_def

run

trait_def

with_tables 🔒

Type Aliases

Opaque

A type that provides internal information but that can still be used for debug purpose.