Module rustc_smir::stable_mir

Module that implements the public interface to the Stable MIR.

This module shall contain all type definitions and APIs that we expect third-party tools to invoke to interact with the compiler.

The goal is to eventually move this module to its own crate which shall be published on crates.io.

Note:

There shouldn’t be any direct references to internal compiler constructs in this module. If you need an internal construct, consider using rustc_internal or rustc_smir.

Modules

mir

ty

Structs

Crate

Holds information about a crate.

CrateItem

Holds information about an item in the crate. For now, it only stores the item DefId. Use functions inside rustc_internal module to use this item.

Statics

TLV 🔒

Traits

Context

Functions

all_local_items

Retrieve all items in the local crate that have a MIR associated with them.

all_trait_decls

all_trait_impls

entry_fn

Return the function where execution starts if the current crate defines that. This is usually main, but could be start if the crate is a no-std crate.

external_crates

Try to find a crate with the given name.

find_crate

Try to find a crate with the given name.

local_crate

Access to the local crate.

run

trait_decl

trait_impl

with 🔒

Loads the current context and calls a function with it. Do not nest these, as that will ICE.

Type Aliases

CrateItems

A list of crate items.

CrateNum

The number that identifies a crate.

DefId

A unique identification number for each item accessible for the current compilation unit.

ImplTraitDecls

A list of impl trait decls.

Symbol

Use String for now but we should replace it.

TraitDecls

A list of trait decls.