Trait rustc_smir::stable_mir::Context

pub trait Context {

    // Required methods
    fn entry_fn(&mut self) -> Option<CrateItem>;
    fn all_local_items(&mut self) -> CrateItems;
    fn mir_body(&mut self, item: &CrateItem) -> Body;
    fn all_trait_decls(&mut self) -> TraitDecls;
    fn trait_decl(&mut self, trait_def: &TraitDef) -> TraitDecl;
    fn all_trait_impls(&mut self) -> ImplTraitDecls;
    fn trait_impl(&mut self, trait_impl: &ImplDef) -> ImplTrait;
    fn generics_of(&mut self, generic_def: &GenericDef) -> Generics;
    fn local_crate(&self) -> Crate;
    fn external_crates(&self) -> Vec<Crate>;
    fn find_crate(&self, name: &str) -> Option<Crate>;
    fn ty_kind(&mut self, ty: Ty) -> TyKind;
    fn rustc_tables(&mut self, f: &mut dyn FnMut(&mut Tables<'_>));

}

Required Methods

fn entry_fn(&mut self) -> Option<CrateItem>

fn all_local_items(&mut self) -> CrateItems

Retrieve all items of the local crate that have a MIR associated with them.

fn mir_body(&mut self, item: &CrateItem) -> Body

fn all_trait_decls(&mut self) -> TraitDecls

fn trait_decl(&mut self, trait_def: &TraitDef) -> TraitDecl

fn all_trait_impls(&mut self) -> ImplTraitDecls

fn trait_impl(&mut self, trait_impl: &ImplDef) -> ImplTrait

fn generics_of(&mut self, generic_def: &GenericDef) -> Generics

fn local_crate(&self) -> Crate

Get information about the local crate.

fn external_crates(&self) -> Vec<Crate>

Retrieve a list of all external crates.

fn find_crate(&self, name: &str) -> Option<Crate>

Find a crate with the given name.

fn ty_kind(&mut self, ty: Ty) -> TyKind

Obtain the representation of a type.

fn rustc_tables(&mut self, f: &mut dyn FnMut(&mut Tables<'_>))

HACK: Until we have fully stable consumers, we need an escape hatch to get DefIds out of CrateItems.