Module rustc_smir::stable_mir::ty

Structs

AdtDef

AliasDef

AliasTy

Allocation

Binder

BoundTy

BrNamedDef

ClosureDef

Const

ConstDef

EarlyBinder

ExistentialProjection

ExistentialTraitRef

FnDef

FnSig

ForeignDef

GeneratorDef

GenericArgs

GenericDef

GenericParamDef

Generics

ImplDef

ParamDef

ParamTy

ProvenanceMap

Stores the provenance information of pointers stored in memory.

TraitDecl

TraitDef

TraitRef

Ty

UnevaluatedConst

Enums

Abi

AliasKind

BoundRegionKind

BoundTyKind

BoundVariableKind

ConstantKind

DynKind

ExistentialPredicate

FloatTy

GenericArgKind

GenericParamDefKind

IntTy

Movability

RigidTy

TermKind

TraitSpecializationKind

TyKind

UintTy

Functions

allocation_filter

Creates an Allocation only from information within the AllocRange.

new_allocation

Type Aliases

Align

Bytes

Ident 🔒

ImplTrait

InitMaskMaterialized

PolyFnSig

Promoted

Prov

Region 🔒

Size

Span