Module rustc_smir::rustc_internal

Module that implements the bridge between Stable MIR and internal compiler MIR.

For that, we define APIs that will temporarily be public to 3P that exposes rustc internal APIs until stable MIR is complete.

Modules

internal 🔒

Module containing the translation from stable mir constructs to the rustc counterpart.

pretty

Structs

IndexMap

Simmilar to rustc’s FxIndexMap, IndexMap with extra safety features added.

Statics

TLV 🔒

Traits

RustcInternal

Trait used to translate a stable construct to its rustc counterpart.

Functions

crate_num

init 🔒

internal

Convert a stable item into its internal Rust compiler counterpart, if one exists.

run

stable

Convert an internal Rust compiler item into its stable counterpart, if one exists.

with_tables 🔒

Loads the current context and calls a function with it. Do not nest these, as that will ICE.