Module rustc_smir::rustc_smir

Module that implements what will become the rustc side of Stable MIR. This module is responsible for building Stable MIR components from internal components.

This module is not intended to be invoked directly by users. It will eventually become the public API of rustc that will be invoked by the stable_mir crate.

For now, we are developing everything inside rustc, thus, we keep this module private.

Modules

alloc 🔒

builder 🔒

Logic required to produce a monomorphic stable body.

context 🔒

Implementation of [stable_mir::compiler_interface::Context] trait.

convert 🔒

Conversion of internal Rust compiler items to stable ones.

Structs

Tables

Traits

Stable

Trait used to convert between an internal MIR type to a Stable MIR type.

Functions

new_item_kind 🔒

smir_crate 🔒

Build a stable mir crate from a given crate number.